move loadRegistry tests to their own file
